move selection-proxy under system
